+++ orders/models/invoice.py
@@ -1,0 +1,18 @@
+from django.db import models
+
+from .order import Order
+
+class Invoice(models.Model):
+    class InvoiceStatus(models.TextChoices):
+        UNPAID = 'UNPAID', 'Unpaid'
+        PAID = 'PAID', 'Paid'
+
+    order = models.OneToOneField(Order, on_delete=models.CASCADE)
+    total_amount = models.DecimalField(max_digits=10, decimal_places=2)
+    status = models.CharField(max_length=10, choices=InvoiceStatus.choices, default=InvoiceStatus.UNPAID)
+    issued_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return f"Invoice for Order {self.order.id}"
+
+
